Cleanup build tasks residue
 - fix(gulp/clean): clean task working all the time
 - feat(gulp/serve): ensure build runs everytime we startup serve
 - refactor(gulp): removing residue, which also fixed the undefined directory
 - chore(gulp): cleaned additional unecessary script boiler plate
 - refactor(gulp): modernized directory structure
