+++ gulpfile.js/tasks/lint.js
