+++ gulpfile.js/tasks/server.js
@@ -1,0 +1,47 @@
+'use strict';
+
+var gulp = require('gulp');
+
+var paths = gulp.paths;
+
+var util = require('util');
+
+var browserSync = require('browser-sync');
+
+function browserSyncInit(baseDir, files, browser) {
+  browser = browser === undefined ? 'default' : browser;
+
+  var routes = null;
+  if(baseDir === paths.src || (util.isArray(baseDir) && baseDir.indexOf(paths.src) !== -1)) {
+    routes = {
+      '/bower_components': 'bower_components'
+    };
+  }
+
+  browserSync.instance = browserSync.init(files, {
+    startPath: '/',
+    server: {
+      baseDir: baseDir,
+      //middleware: middleware,
+      routes: routes
+    },
+    open: false
+  });
+}
+
+gulp.task('serve', ['watch'], function () {
+  browserSyncInit([
+    paths.tmp + '/serve',
+    paths.tmp + '/partials',
+    paths.src
+  ], [
+    paths.tmp + '/serve/**/*.css',
+    paths.tmp + '/partials/**/*.js',
+    paths.src + '/**/*.js',
+    paths.src + 'src/assets/images/**/*',
+    paths.tmp + '/serve/*.html',
+    paths.tmp + '/serve/**/*.js',
+    paths.tmp + '/serve/**/*.html',
+    paths.src + '/**/*.html'
+  ]);
+});
